+++ NEWS.md
@@ -1,5 +1,7 @@
 # lifecycle (development version)
 
+* The `what` and `with` arguments no longer need to be quoted with `""`.
+
 * A character vector `details` is now converted into a bulleted list (#55).
 
 * You can now deprecate an argument with `foo(arg)` instead of `foo(arg =)` (#78).